+++ ext/php_cassandra.c
@@ -30,8 +30,8 @@
 #define PHP_CASSANDRA_DEFAULT_LOG       "cassandra.log"
 #define PHP_CASSANDRA_DEFAULT_LOG_LEVEL "ERROR"
 
-#if CURRENT_CPP_DRIVER_VERSION < CPP_DRIVER_VERSION(2, 2, 2)
-#error C/C++ driver version 2.2.2 or greater required
+#if CURRENT_CPP_DRIVER_VERSION < CPP_DRIVER_VERSION(2, 3, 0)
+#error C/C++ driver version 2.3.0 or greater required
 #endif
 
 static uv_once_t log_once = UV_ONCE_INIT;
@@ -273,7 +273,8 @@ static PHP_GINIT_FUNCTION(cassandra)
 {
   uv_once(&log_once, php_cassandra_log_initialize);
 
-  cassandra_globals->uuid_gen            = cass_uuid_gen_new();
+  cassandra_globals->uuid_gen            = NULL;
+  cassandra_globals->uuid_gen_pid        = 0;
   cassandra_globals->persistent_clusters = 0;
   cassandra_globals->persistent_sessions = 0;
   PHP5TO7_ZVAL_UNDEF(cassandra_globals->type_varchar);
@@ -281,6 +282,7 @@ static PHP_GINIT_FUNCTION(cassandra)
   PHP5TO7_ZVAL_UNDEF(cassandra_globals->type_blob);
   PHP5TO7_ZVAL_UNDEF(cassandra_globals->type_ascii);
   PHP5TO7_ZVAL_UNDEF(cassandra_globals->type_bigint);
+  PHP5TO7_ZVAL_UNDEF(cassandra_globals->type_smallint);
   PHP5TO7_ZVAL_UNDEF(cassandra_globals->type_counter);
   PHP5TO7_ZVAL_UNDEF(cassandra_globals->type_int);
   PHP5TO7_ZVAL_UNDEF(cassandra_globals->type_varint);
@@ -296,7 +298,9 @@ static PHP_GINIT_FUNCTION(cassandra)
 
 static PHP_GSHUTDOWN_FUNCTION(cassandra)
 {
-  cass_uuid_gen_free(cassandra_globals->uuid_gen);
+  if (cassandra_globals->uuid_gen) {
+    cass_uuid_gen_free(cassandra_globals->uuid_gen);
+  }
   php_cassandra_log_cleanup();
 }
 
@@ -342,11 +346,15 @@ PHP_MINIT_FUNCTION(cassandra)
   cassandra_define_Value(TSRMLS_C);
   cassandra_define_Numeric(TSRMLS_C);
   cassandra_define_Bigint(TSRMLS_C);
+  cassandra_define_Smallint(TSRMLS_C);
+  cassandra_define_Tinyint(TSRMLS_C);
   cassandra_define_Blob(TSRMLS_C);
   cassandra_define_Decimal(TSRMLS_C);
   cassandra_define_Float(TSRMLS_C);
   cassandra_define_Inet(TSRMLS_C);
   cassandra_define_Timestamp(TSRMLS_C);
+  cassandra_define_Date(TSRMLS_C);
+  cassandra_define_Time(TSRMLS_C);
   cassandra_define_UuidInterface(TSRMLS_C);
   cassandra_define_Timeuuid(TSRMLS_C);
   cassandra_define_Uuid(TSRMLS_C);
@@ -387,6 +395,14 @@ PHP_MINIT_FUNCTION(cassandra)
   cassandra_define_DefaultTable(TSRMLS_C);
   cassandra_define_Column(TSRMLS_C);
   cassandra_define_DefaultColumn(TSRMLS_C);
+  cassandra_define_Index(TSRMLS_C);
+  cassandra_define_DefaultIndex(TSRMLS_C);
+  cassandra_define_MaterializedView(TSRMLS_C);
+  cassandra_define_DefaultMaterializedView(TSRMLS_C);
+  cassandra_define_Function(TSRMLS_C);
+  cassandra_define_DefaultFunction(TSRMLS_C);
+  cassandra_define_Aggregate(TSRMLS_C);
+  cassandra_define_DefaultAggregate(TSRMLS_C);
 
   cassandra_define_Type(TSRMLS_C);
   cassandra_define_TypeScalar(TSRMLS_C);
@@ -419,6 +435,12 @@ PHP_MSHUTDOWN_FUNCTION(cassandra)
 
 PHP_RINIT_FUNCTION(cassandra)
 {
+#define XX_SCALAR(name, value) \
+  PHP5TO7_ZVAL_UNDEF(CASSANDRA_G(type_##name));
+
+  PHP_CASSANDRA_SCALAR_TYPES_MAP(XX_SCALAR)
+#undef XX_SCALAR
+
   return SUCCESS;
 }
 
@@ -565,9 +587,13 @@ throw_invalid_argument(zval *object,
                               "%s must be %s, an instance of Unknown Class given",
                               object_name, expected_type);
     }
-  } else {
+  } else if (Z_TYPE_P(object) == IS_STRING) {
     zend_throw_exception_ex(cassandra_invalid_argument_exception_ce, 0 TSRMLS_CC,
                             "%s must be %s, '%Z' given",
+                            object_name, expected_type, object);
+  } else {
+    zend_throw_exception_ex(cassandra_invalid_argument_exception_ce, 0 TSRMLS_CC,
+                            "%s must be %s, %Z given",
                             object_name, expected_type, object);
   }
 }