+++ ext/php_cassandra.h
@@ -38,6 +38,19 @@
 #include <Zend/zend_exceptions.h>
 #include <Zend/zend_interfaces.h>
 
+#ifdef HAVE_SYS_TYPES_H
+#include <sys/types.h>
+#endif
+
+#ifdef HAVE_UNISTD_H
+#include <unistd.h>
+#endif
+
+#ifdef PHP_WIN32
+typedef int pid_t;
+#include <process.h>
+#endif
+
 #if PHP_VERSION_ID < 50304
 #  error PHP 5.3.4 or later is required in order to build the driver
 #endif
@@ -116,7 +129,8 @@ php5to7_string_compare(php5to7_string s1, php5to7_string s2)
   php_cassandra_##type_name##_object_fetch(object);
 
 #define PHP5TO7_SMART_STR_INIT { NULL, 0 }
-#define PHP5TO7_SMART_STR_VAL(ss) (ss).s->val
+#define PHP5TO7_SMART_STR_VAL(ss) ((ss).s ? (ss).s->val : NULL)
+#define PHP5TO7_SMART_STR_LEN(ss) ((ss).s ? (ss).s->len : 0)
 
 #define PHP5TO7_STRCMP(s, c) strcmp((s)->val, (c))
 #define PHP5TO7_STRVAL(s) ((s)->val)
@@ -143,6 +157,12 @@ php5to7_string_compare(php5to7_string s1, php5to7_string s2)
 
 #define PHP5TO7_ADD_ASSOC_ZVAL_EX(zv, key, len, val) \
   add_assoc_zval_ex((zv), (key), (size_t)(len - 1), val)
+
+#define PHP5TO7_ADD_ASSOC_STRINGL_EX(zv, key, key_len, str, str_len) \
+  add_assoc_stringl_ex((zv), (key), (size_t)(key_len - 1), (char *)(str), (size_t)(str_len))
+
+#define PHP5TO7_ADD_NEXT_INDEX_STRING(zv, str) \
+  add_next_index_string((zv), (char*)(str));
 
 #define PHP5TO7_ZEND_HASH_FOREACH_VAL(ht, _val) \
   ZEND_HASH_FOREACH_VAL(ht, _val)
@@ -228,6 +248,7 @@ php5to7_string_compare(php5to7_string s1, php5to7_string s2)
 #define PHP5TO7_ZVAL_STRING(zv, s) ZVAL_STRING(zv, s)
 #define PHP5TO7_ZVAL_STRINGL(zv, s, len) ZVAL_STRINGL(zv, s, len)
 #define PHP5TO7_RETVAL_STRING(s) RETVAL_STRING(s)
+#define PHP5TO7_RETURN_STRING(s) RETURN_STRING(s)
 #define PHP5TO7_RETVAL_STRINGL(s, len) RETVAL_STRINGL(s, len)
 #define PHP5TO7_RETURN_STRINGL(s, len) RETURN_STRINGL(s, len)
 
@@ -271,6 +292,7 @@ php5to7_string_compare(php5to7_string s1, php5to7_string s2)
 
 #define PHP5TO7_SMART_STR_INIT { NULL, 0, 0 }
 #define PHP5TO7_SMART_STR_VAL(ss) (ss).c
+#define PHP5TO7_SMART_STR_LEN(ss) (ss).len
 
 #define PHP5TO7_STRCMP(s, c) strcmp((s), (c))
 #define PHP5TO7_STRVAL(s) (s)
@@ -298,6 +320,12 @@ php5to7_string_compare(php5to7_string s1, php5to7_string s2)
 
 #define PHP5TO7_ADD_ASSOC_ZVAL_EX(zv, key, len, val) \
   add_assoc_zval_ex((zv), (key), (uint)(len), val)
+
+#define PHP5TO7_ADD_ASSOC_STRINGL_EX(zv, key, key_len, str, str_len) \
+  add_assoc_stringl_ex((zv), (key), (uint)(key_len), (char*)(str), (uint)(str_len), 1)
+
+#define PHP5TO7_ADD_NEXT_INDEX_STRING(zv, str) \
+  add_next_index_string((zv), (char*)(str), 1);
 
 #define PHP5TO7_ZEND_HASH_FOREACH_VAL(ht, _val) do { \
   HashPosition _pos; \
@@ -397,6 +425,7 @@ php5to7_string_compare(php5to7_string s1, php5to7_string s2)
 #define PHP5TO7_ZVAL_STRING(zv, s) ZVAL_STRING(zv, s, 1)
 #define PHP5TO7_ZVAL_STRINGL(zv, s, len) ZVAL_STRINGL(zv, s, len, 1)
 #define PHP5TO7_RETVAL_STRING(s) RETVAL_STRING(s, 1)
+#define PHP5TO7_RETURN_STRING(s) RETURN_STRING(s, 1)
 #define PHP5TO7_RETVAL_STRINGL(s, len) RETVAL_STRINGL(s, len, 1)
 #define PHP5TO7_RETURN_STRINGL(s, len) RETURN_STRINGL(s, len, 1)
 
@@ -462,6 +491,7 @@ PHP_MINFO_FUNCTION(cassandra);
 
 ZEND_BEGIN_MODULE_GLOBALS(cassandra)
   CassUuidGen          *uuid_gen;
+  pid_t                 uuid_gen_pid;
   unsigned int          persistent_clusters;
   unsigned int          persistent_sessions;
   php5to7_zval          type_varchar;
@@ -478,8 +508,12 @@ ZEND_BEGIN_MODULE_GLOBALS(cassandra)
   php5to7_zval          type_float;
   php5to7_zval          type_inet;
   php5to7_zval          type_timestamp;
+  php5to7_zval          type_date;
+  php5to7_zval          type_time;
   php5to7_zval          type_uuid;
   php5to7_zval          type_timeuuid;
+  php5to7_zval          type_smallint;
+  php5to7_zval          type_tinyint;
 ZEND_END_MODULE_GLOBALS(cassandra)
 
 #ifdef ZTS