+++ ext/src/Cassandra/Cluster/Builder.c
@@ -50,7 +50,7 @@ PHP_METHOD(ClusterBuilder, build)
     php5to7_zend_resource_le *le;
 
     hash_key_len = spprintf(&hash_key, 0,
-      "cassandra:%s:%d:%d:%s:%d:%d:%d:%s:%s:%d:%d:%d:%d:%d:%d:%d:%d:%d:%d:%d:%d",
+      "cassandra:%s:%d:%d:%s:%d:%d:%d:%s:%s:%d:%d:%d:%d:%d:%d:%d:%d:%d:%d:%d:%d:%s:%s:%s:%s",
       builder->contact_points, builder->port, builder->load_balancing_policy,
       SAFE_STR(builder->local_dc), builder->used_hosts_per_remote_dc,
       builder->allow_remote_dcs_for_local_cl, builder->use_token_aware_routing,
@@ -60,7 +60,9 @@ PHP_METHOD(ClusterBuilder, build)
       builder->core_connections_per_host, builder->max_connections_per_host,
       builder->reconnect_interval, builder->enable_latency_aware_routing,
       builder->enable_tcp_nodelay, builder->enable_tcp_keepalive,
-      builder->tcp_keepalive_delay, builder->enable_schema);
+      builder->tcp_keepalive_delay, builder->enable_schema,
+      SAFE_STR(builder->whitelist_hosts), SAFE_STR(builder->whitelist_dcs),
+      SAFE_STR(builder->blacklist_hosts), SAFE_STR(builder->blacklist_dcs));
 
     cluster->hash_key     = hash_key;
     cluster->hash_key_len = hash_key_len;
@@ -657,11 +659,10 @@ PHP_METHOD(ClusterBuilder, withProtocolVersion)
   builder = PHP_CASSANDRA_GET_CLUSTER_BUILDER(getThis());
 
   if (Z_TYPE_P(version) == IS_LONG &&
-      Z_LVAL_P(version) < 3 &&
-      Z_LVAL_P(version) > 0) {
+      Z_LVAL_P(version) >= 1) {
     builder->protocol_version = Z_LVAL_P(version);
   } else {
-    INVALID_ARGUMENT(version, "either 1 or 2");
+    INVALID_ARGUMENT(version, "must be >= 1");
   }
 
   RETURN_ZVAL(getThis(), 1, 0);
@@ -1144,29 +1145,29 @@ php_cassandra_cluster_builder_properties(zval *object TSRMLS_DC)
   }
 
   PHP5TO7_ZVAL_MAYBE_MAKE(blacklistHosts);
-  if (!PHP5TO7_ZVAL_IS_UNDEF(self->blacklist_hosts)) {
-    PHP5TO7_ZVAL_COPY(PHP5TO7_ZVAL_MAYBE_P(blacklistHosts), PHP5TO7_ZVAL_MAYBE_P(self->blacklist_hosts));
+  if (self->blacklist_hosts) {
+    PHP5TO7_ZVAL_STRING(PHP5TO7_ZVAL_MAYBE_P(blacklistHosts), self->blacklist_hosts);
   } else {
     ZVAL_NULL(PHP5TO7_ZVAL_MAYBE_P(blacklistHosts));
   }
 
   PHP5TO7_ZVAL_MAYBE_MAKE(whitelistHosts);
-  if (!PHP5TO7_ZVAL_IS_UNDEF(self->whitelist_hosts)) {
-    PHP5TO7_ZVAL_COPY(PHP5TO7_ZVAL_MAYBE_P(whitelistHosts), PHP5TO7_ZVAL_MAYBE_P(self->whitelist_hosts));
+  if (self->whitelist_hosts) {
+    PHP5TO7_ZVAL_STRING(PHP5TO7_ZVAL_MAYBE_P(whitelistHosts), self->whitelist_hosts);
   } else {
     ZVAL_NULL(PHP5TO7_ZVAL_MAYBE_P(whitelistHosts));
   }
 
   PHP5TO7_ZVAL_MAYBE_MAKE(blacklistDCs);
-  if (!PHP5TO7_ZVAL_IS_UNDEF(self->blacklist_dcs)) {
-    PHP5TO7_ZVAL_COPY(PHP5TO7_ZVAL_MAYBE_P(blacklistDCs), PHP5TO7_ZVAL_MAYBE_P(self->blacklist_dcs));
+  if (self->blacklist_dcs) {
+    PHP5TO7_ZVAL_STRING(PHP5TO7_ZVAL_MAYBE_P(blacklistDCs), self->blacklist_dcs);
   } else {
     ZVAL_NULL(PHP5TO7_ZVAL_MAYBE_P(blacklistDCs));
   }
 
   PHP5TO7_ZVAL_MAYBE_MAKE(whitelistDCs);
-  if (!PHP5TO7_ZVAL_IS_UNDEF(self->whitelist_dcs)) {
-    PHP5TO7_ZVAL_COPY(PHP5TO7_ZVAL_MAYBE_P(whitelistDCs), PHP5TO7_ZVAL_MAYBE_P(self->whitelist_dcs));
+  if (self->whitelist_dcs) {
+    PHP5TO7_ZVAL_STRING(PHP5TO7_ZVAL_MAYBE_P(whitelistDCs), self->whitelist_dcs);
   } else {
     ZVAL_NULL(PHP5TO7_ZVAL_MAYBE_P(whitelistDCs));
   }
@@ -1278,6 +1279,26 @@ php_cassandra_cluster_builder_free(php5to7_zend_object_free *object TSRMLS_DC)
     self->password = NULL;
   }
 
+  if (self->whitelist_hosts) {
+    efree(self->whitelist_hosts);
+    self->whitelist_hosts = NULL;
+  }
+
+  if (self->blacklist_hosts) {
+    efree(self->blacklist_hosts);
+    self->blacklist_hosts = NULL;
+  }
+
+  if (self->whitelist_dcs) {
+    efree(self->whitelist_dcs);
+    self->whitelist_dcs = NULL;
+  }
+
+  if (self->blacklist_dcs) {
+    efree(self->blacklist_dcs);
+    self->whitelist_dcs = NULL;
+  }
+
   PHP5TO7_ZVAL_MAYBE_DESTROY(self->ssl_options);
   PHP5TO7_ZVAL_MAYBE_DESTROY(self->default_timeout);
   PHP5TO7_ZVAL_MAYBE_DESTROY(self->retry_policy);
@@ -1307,7 +1328,7 @@ php_cassandra_cluster_builder_new(zend_class_entry *ce TSRMLS_DC)
   self->default_consistency = PHP_CASSANDRA_DEFAULT_CONSISTENCY;
   self->default_page_size = 5000;
   self->persist = 1;
-  self->protocol_version = 3;
+  self->protocol_version = 4;
   self->io_threads = 1;
   self->core_connections_per_host = 1;
   self->max_connections_per_host = 2;